+++ R/get_contrast_data_character.R
@@ -22,8 +22,19 @@ get_contrast_data_character <- function(model,
     if (isTRUE(flag)) {
         levs_idx <- contrast_categories_shortcuts(levs, variable, interaction)
 
+    # manual data frame
+    } else if (isTRUE(checkmate::check_data_frame(variable$value))) {
+        levs_idx <- contrast_categories_df(variable)
+        lab <- "manual"
+
+    # custom function
+    } else if (isTRUE(checkmate::check_function(variable$value))) {
+        tmp <- variable$value(x)
+        levs_idx <- data.table::data.table(lo = tmp[, 1], hi = tmp[, 2])
+        lab <- "custom"
+
     # vector of two values
-    } else if (length(variable$value) == 2) {
+    } else if (isTRUE(checkmate::check_atomic_vector(variable$value, len = 2))) {
         if (is.character(variable$value)) {
             tmp <- modeldata[[variable$name]]
             if (any(!variable$value %in% as.character(tmp))) {