+++ R/plot_comparisons.R
@@ -168,4 +168,7 @@ plot_comparisons <- function(model,
 #' @inherit plot_predictions
 #' @keywords internal
 #' @export
-plot_cco <- plot_comparisons
+plot_cco <- function(...) {
+    insight::format_warning("This function has been renamed to `plot_comparisons()`. The `plot_cco()` alias will be removed in a future release.")
+    plot_comparisons(...)
+}